Tools: Move device open/close into device module.
